requirements.txt
py-cpuinfo
psutil
InquirerPy
py-cpuinfo
